GH-517 Move ability profile to first tab
