+++ templates/feedback/learningprogress.mustache
@@ -32,14 +32,7 @@
                 "charttitle": "Person ability relative to parentscale",
                 "chart": "<div>...</div>"
             }
-        ],
-        "abilityprofile": [
-            {
-                "charttitle": "Person ability relative to parentscale",
-                "chart": "<div>...</div>"
-            }
-        ]
-    }
+        ]    }
 }}
 {{#progressindividual}}
     <h6>
@@ -53,9 +46,3 @@
     </h6>
     {{{chart}}}
 {{/progresscomparison}}
-{{#abilityprofile}}
-    <h6>
-    {{charttitle}}
-    </h6>
-    {{{chart}}}
-{{/abilityprofile}}
